+++ server/src/TcpServer.cpp
@@ -2,7 +2,7 @@
 #include "MessageRouter.h"
 #include <iostream>
 #include <istream>
-#include <boost/json/src.hpp>
+#include <boost/json.hpp>
 
 #include "ConnectionManager.h"