+++ changelog.md
@@ -1,1 +1,1 @@
-* Remove non header only boost libraries so its not required to be installed on system
+* Implement Ai Tutor on client and serverside, and some small fixes